Moved checkErrorWrapper to utils
